+++ Console_Text_RPG_Team/Console_Text_RPG_Team/SceneBattleAttack.cs
@@ -118,11 +118,11 @@ namespace Console_Text_RPG_Team
             }
         }
 
-        public float SelectDamage(int result, Player player)
+        public int SelectDamage(int result, Player player)
         {
             while (true)
             {
-                float damage = 0;
+                int damage = 0;
                 if (result == 2)
                 {
 
@@ -175,10 +175,10 @@ namespace Console_Text_RPG_Team
                     }
 
                     (int result, _player) = WhatSelectDamage(_player);
-                    float damaged = SelectDamage(result, _player);
+                    int damaged = SelectDamage(result, _player);
                     if (damaged == 0) continue; // 물약 사용 시 턴 종료
-                    float criticalDamage = _player.CriticalDamage(_player, damaged);
-                    float finalDamage = GetRandomDamage(criticalDamage);
+                    int criticalDamage = _player.CriticalDamage(_player, damaged);
+                    int finalDamage = GetRandomDamage(criticalDamage);
                     target.TakeDamage(finalDamage);
                     PlayerAttackLog(_player, target, finalDamage);
                     break; // 공격 후 플레이어 턴 종료
@@ -199,7 +199,7 @@ namespace Console_Text_RPG_Team
             {
                 if (!monster.IsAlive()) continue;
 
-                float damage = GetRandomDamage(monster.atk);
+                int damage = GetRandomDamage(monster.atk);
                 player.TakeDamage(damage);
                 MonsterAttackLog(monster, player, damage);
             }
@@ -234,9 +234,9 @@ namespace Console_Text_RPG_Team
 
         private Random rand = new Random();
 
-        public float GetRandomDamage(float defaultDamage)
+        public int GetRandomDamage(int defaultDamage)
         {
-            float error = defaultDamage * 0.1f;
+            int error = (int)(defaultDamage * 0.1f);
             int roundedError = (int)(error + 0.999f); // 올림 작업
 
             int minDamage = (int)(defaultDamage - roundedError);
@@ -277,7 +277,7 @@ namespace Console_Text_RPG_Team
             Console.Write(" >>");
         }
 
-        public void PlayerAttackLog(Player attacker, Monster target, float damage)
+        public void PlayerAttackLog(Player attacker, Monster target, int damage)
         {
             Console.Clear();
             StringBuilder sb = new StringBuilder();
@@ -311,7 +311,7 @@ namespace Console_Text_RPG_Team
 
         }
 
-		public void MonsterAttackLog(Monster attacker, Player target, float damage)
+		public void MonsterAttackLog(Monster attacker, Player target, int damage)
 		{
 			StringBuilder sb = new StringBuilder();
 
@@ -397,7 +397,7 @@ namespace Console_Text_RPG_Team
             }
             int prevLevel = player.level;
             int prevExp = player.exp;
-            float prevHP = player.maxHp;
+            int prevHP = player.maxHp;
             int prevGold = player.Gold;
 
             player.Gold += totalGold;
